+++ .gitignore
@@ -68,3 +68,4 @@ docs/_build
 .envrc
 *.ipynb
 tests/test_parsed_engines.py
+tests/test_engines.py
